migration islemleri alembic ogrendik ve kullandik

+++ ToDoGemini/routers/todo.py
@@ -7,7 +7,7 @@ from sqlalchemy.orm import Session
 from database import engine, SessionLocal
 #database baglantisi sagliyoruz sessionlocal kullanarak
 from typing import Annotated
-from routers.auth import router as auth_router
+from routers.auth import get_current_user
 
 
 router=APIRouter(
@@ -33,29 +33,35 @@ def get_db():
         db.close()
         #islem bittiginde baglantyi kapatmak icni kullaniyorz
 
-
 db_dependency=Annotated[Session, Depends(get_db)]
 #veri tabani icin baglanti sagliyor ve surekli ayni seyi yazmak yerine degisken olusturup bu sekilde kullanmak daha mantikli
+user_dependency=Annotated[dict, Depends(get_current_user)]
 
 
 
-@router.get("/read_all")
-async def read_all(db: db_dependency):
-    return db.query(Todo).all()
+@router.get("/")
+async def read_all(user: user_dependency, db: db_dependency):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED)
+    return db.query(Todo).filter(Todo.owner_id==user.get('id')).all()
 
 
-@router.get("/get_by_id/{todo_id}", status_code=status.HTTP_200_OK)
-async def read_by_id(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo = db.query(Todo).filter(Todo.id == todo_id).first()
+@router.get("/{todo_id}", status_code=status.HTTP_200_OK)
+async def read_by_id(user:user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED)
+    todo = db.query(Todo).filter(Todo.id == todo_id).filter(Todo.owner_id==user.get('id')).first()
     # tek bir todo vermiyor olurda birden fazla todo geliyorsa liste donduruyor ilk id aliyor ondan
     if todo is not None:
         return todo
     raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Todo not found")
 
 
-@router.post("/create_todo", status_code=status.HTTP_201_CREATED)
-async def create_todo(db: db_dependency, todo_request: TodoRequest):
-    todo = Todo(**todo_request.dict())
+@router.post("/", status_code=status.HTTP_201_CREATED)
+async def create_todo(user:user_dependency, db: db_dependency, todo_request: TodoRequest):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED)
+    todo = Todo(**todo_request.dict(), owner_id=user.get('id'))
     db.add(todo)
     db.commit()
     db.refresh(todo)  # ID gibi alanları günceller
@@ -63,11 +69,14 @@ async def create_todo(db: db_dependency, todo_request: TodoRequest):
     #db.commit islemin yapilacagi anlamina geliyor eger eklemezseniz islem yapilmaz
 
 
-@router.put("/update_todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async  def update_todo(db: db_dependency,
+@router.put("/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
+async  def update_todo(user: user_dependency,
+                       db: db_dependency,
                        todo_request: TodoRequest,
                        todo_id : int =Path(gt=0)):
-    todo=db.query(Todo).filter(Todo.id==todo_id).first()
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED)
+    todo = db.query(Todo).filter(Todo.id == todo_id).filter(Todo.owner_id == user.get('id')).first()
     if todo is None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Todo not found")
     todo.title=todo_request.title
@@ -77,9 +86,11 @@ async  def update_todo(db: db_dependency,
     db.add(todo)
     db.commit()
 
-@router.delete("/delete_todo/{todo_id}",status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db:db_dependency, todo_id:int = Path(gt=0)):
-    todo=db.query(Todo).filter(Todo.id==todo_id).first
+@router.delete("/{todo_id}",status_code=status.HTTP_204_NO_CONTENT)
+async def delete_todo(user:user_dependency, db:db_dependency, todo_id:int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED)
+    todo = db.query(Todo).filter(Todo.id == todo_id).filter(Todo.owner_id == user.get('id')).first()
     if todo is None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Todo not found")
     db.query(Todo).filter(Todo.id==todo_id).delete()
